requirements.txt
opencv-python
numpy
matplotlib
ultralytics
torch
scipy
python-chess
pandas
ipython
streamlit
scikit_learn